Ajustes na lógica de início e de jogo. Ajustes nas exportações
