+++ logicaDeVisualizacao.js
@@ -1,3 +1,5 @@
+const prompt = require("prompt-sync")();
+
 function geraTabuleiro(){
   let tabuleiro = []
   for (let i = 0; i < 5; i++) {
@@ -8,6 +10,27 @@ function geraTabuleiro(){
   }
   return tabuleiro
 }
+
+function apresentacao() {
+  console.log("\n         ..................................." +
+  "\n           ### Jogo de Batalha Naval ####" +
+  "\n         ..................................." +
+  "\n         Jogo de tabuleiro de dois jogadores" +
+  "\nSeu objectivo é derrubar os barcos do oponente adversário," +
+  "\nganha quem derrubar todos os navios adversários primeiro." +
+  "\n")
+  console.log("!! Vamos Jogar !!")
+  console.log("Qualquer caracter = sim")
+  let jogar = prompt("Ou so enter para nao")
+  if (!jogar == "") {
+    console.clear()
+    return true
+  } else {
+    console.log(" \n !! Que pena !!")
+    return false
+  }
+}
+
 
 function logicaVisualizacao() { // RenatoTonelli 22jun1647 - funcao criada para ser chamado no jogo
   let tabuleiro = geraTabuleiro()
@@ -36,7 +59,8 @@ function logicaVisualizacao() { // RenatoTonelli 22jun1647 - funcao criada para 
 // Por RenatoTonelli 22jun-1651 - criando uma simples chamada para apresentacao do jogo
 module.exports = {
   "apresentaLogica":logicaVisualizacao,
-  "geraTabuleiro": geraTabuleiro
+  "geraTabuleiro": geraTabuleiro,
+  "apresentacao": apresentacao
 }
